Fix: 나이 타입 변경

+++ core-model/src/main/java/com/hmoa/core_model/response/PerfumeAgeResponseDto.kt
@@ -4,6 +4,6 @@ import kotlinx.serialization.Serializable
 
 @Serializable
 data class PerfumeAgeResponseDto(
-    var age: Int,
+    var age: Float,
     var writed: Boolean
 )
